Reorganized client code
