+++ src/app/app-routing.module.ts
@@ -1,6 +1,6 @@
 import {NgModule} from '@angular/core';
 import {RouterModule, Routes} from '@angular/router';
-import {HomeComponent} from './components/home/home.component';
+import {HomeComponent} from './screens/home/home.component';
 
 const routes: Routes = [
   {
