+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import {Component} from '@angular/core';
-import {ElectronService} from './services/electron.service';
+import {ElectronService} from './common/services/electron.service';
 import {TranslateService} from '@ngx-translate/core';
 import {AppConfig} from '../environments/environment';
