+++ src/app/app.module.ts
@@ -12,12 +12,12 @@ import {AppRoutingModule} from './app-routing.module';
 import {TranslateLoader, TranslateModule} from '@ngx-translate/core';
 import {TranslateHttpLoader} from '@ngx-translate/http-loader';
 
-import {ElectronService} from './services/electron.service';
+import {ElectronService} from './common/services/electron.service';
 
 import {WebviewDirective} from './directives/webview.directive';
 
 import {AppComponent} from './app.component';
-import {HomeComponent} from './components/home/home.component';
+import {HomeComponent} from './screens/home/home.component';
 import {SettingsComponent} from './components/settings/settings.component';
 import {SettingsService} from './services/settings.service';
 import {LoadingSpinnerComponent} from './components/loading-spinner/loading-spinner.component';
@@ -25,24 +25,24 @@ import {RepoViewComponent} from './components/repo-view/repo-view.component';
 import {ChangeListComponent} from './components/change-list/change-list.component';
 import {DiffViewerComponent} from './components/diff-viewer/diff-viewer.component';
 import {CommitHistoryComponent} from './components/commit-history/commit-history.component';
-import {ErrorMessageComponent} from './components/common/components/error-message/error-message.component';
+import {ErrorMessageComponent} from './common/components/error-message/error-message.component';
 import {NewTabPageComponent} from './components/new-tab-page/new-tab-page.component';
-import {FilterObjectPipe, FilterPipe} from './components/common/pipes/filter.pipe';
+import {FilterObjectPipe, FilterPipe} from './common/pipes/filter.pipe';
 import {BranchTreeItemComponent} from './components/branch-tree-item/branch-tree-item.component';
-import {FileInputComponent} from './components/common/components/file-input/file-input.component';
-import {GlobalErrorHandlerService} from './components/common/services/global-error-handler.service';
+import {FileInputComponent} from './common/components/file-input/file-input.component';
+import {GlobalErrorHandlerService} from './common/services/global-error-handler.service';
 import {InfiniteScrollModule} from 'ngx-infinite-scroll';
-import {InputModalComponent} from './components/common/components/input-modal/input-modal.component';
+import {InputModalComponent} from './common/components/input-modal/input-modal.component';
 import {HighlightModule} from 'ngx-highlightjs';
 import {GitConfigComponent} from './components/git-config/git-config.component';
-import {AutofocusDirective} from './components/common/directives/autofocus.directive';
+import {AutofocusDirective} from './common/directives/autofocus.directive';
 import {CodeWatcherAlertsComponent} from './components/code-watcher-alerts/code-watcher-alerts.component';
 import {AddWorktreeComponent} from './components/add-worktree/add-worktree.component';
 import {CloneComponent} from './components/clone/clone.component';
 import {NgbModule} from '@ng-bootstrap/ng-bootstrap';
 import {AddSubmoduleComponent} from './components/add-submodule/add-submodule.component';
 import {GitGraphCanvasComponent} from './components/git-graph-canvas/git-graph-canvas.component';
-import {ModalComponent} from './components/common/components/modal/modal.component';
+import {ModalComponent} from './common/components/modal/modal.component';
 import { ConfirmModalComponent } from './components/confirm-modal/confirm-modal.component';
 
 // AoT requires an exported function for factories
