+++ src/app/common/components/input-modal/input-modal.component.scss
