+++ src/app/components/branch-tree-item/branch-tree-item.component.ts
@@ -2,7 +2,7 @@ import {Component, EventEmitter, Input, OnInit, Output} from '@angular/core';
 import {BranchModel} from '../../../../shared/git/Branch.model';
 import {WorktreeModel} from '../../../../shared/git/worktree.model';
 import {SettingsService} from '../../services/settings.service';
-import {FilterPipe} from '../common/pipes/filter.pipe';
+import {FilterPipe} from '../../common/pipes/filter.pipe';
 
 @Component({
   selector: 'app-branch-tree-item',
