+++ src/app/components/clone/clone.component.ts
@@ -1,7 +1,7 @@
 import {ApplicationRef, Component, EventEmitter, OnInit, Output} from '@angular/core';
 import {Observable} from 'rxjs';
 import {GitService} from '../../services/git.service';
-import {ModalService} from '../../services/modal.service';
+import {ModalService} from '../../common/services/modal.service';
 
 @Component({
   selector: 'app-clone',
