+++ src/app/components/code-watcher-alerts/code-watcher-alerts.component.ts
@@ -3,7 +3,7 @@ import {DiffHeaderModel} from '../../../../shared/git/diff.header.model';
 import {CodeWatcherModel} from '../../../../shared/code-watcher.model';
 import {GitService} from '../../services/git.service';
 import {SettingsService} from '../../services/settings.service';
-import {ErrorService} from '../common/services/error.service';
+import {ErrorService} from '../../common/services/error.service';
 import {ErrorModel} from '../../../../shared/common/error.model';
 import {DiffHunkModel} from '../../../../shared/git/diff.hunk.model';
 import {CodeWatcherService, ShowWatchersRequest} from '../../services/code-watcher.service';
