+++ src/app/screens/home/home.component.ts
@@ -1,11 +1,11 @@
 import {ApplicationRef, Component, OnInit} from '@angular/core';
-import {ElectronService} from '../../services/electron.service';
+import {ElectronService} from '../../common/services/electron.service';
 import {HttpClient} from '@angular/common/http';
 import {SettingsService} from '../../services/settings.service';
 import {RepositoryModel} from '../../../../shared/git/Repository.model';
 import {GitService} from '../../services/git.service';
 import {ErrorModel} from '../../../../shared/common/error.model';
-import {ErrorService} from '../common/services/error.service';
+import {ErrorService} from '../../common/services/error.service';
 
 @Component({
   selector: 'app-home',
